move args to separate file
