+++ athena/bin/arguments.py
@@ -1,0 +1,107 @@
+import argparse
+
+
+def parse_args():
+    """
+    Parse cmd line arguments
+
+    Returns
+    -------
+    argparse.Namespace
+        parsed command line arguments
+    """
+
+    parser = argparse.ArgumentParser(
+        description='Generate coverage stats and HTML report.'
+    )
+    parser = generic_arguments(parser)
+    parser = subParsers(parser).parser
+
+    return parser.parse_args()
+
+
+def generic_arguments(parser):
+    """
+    Generic arguments not specific to any running mode
+    """
+    parser.add_argument(
+        '--output', '-o', required=False,
+        help=(
+            'prefix for naming output files, if not given will default to '
+            'using prefix of input sample file.'
+        )
+    )
+
+    return parser
+
+
+class subParsers():
+    """
+    Sub parsers for running individual sub commands
+    """
+    def __init__(self, parser) -> None:
+        self.parser = parser
+        self.subparsers = self.parser.add_subparsers(
+            title='sub_command', dest='sub',
+            help='sub-commands for development and testing'
+        )
+        self.add_annotated_bed_file()
+        self.add_calculate_sample_stats()
+
+
+    def add_annotated_bed_file(self):
+        """
+        _summary_
+        """
+        annotate_parser = self.subparsers.add_parser(
+            'annotate_bed_file',
+            help=(
+                'annotate target bed file with transcript-exon information '
+                'and per base coverage values'
+            )
+        )
+        annotate_parser.add_argument(
+            '--target_bed',
+            help='bed file of target / panel to annotate'
+        )
+        annotate_parser.add_argument(
+            '--exon data',
+            help='tsv file of transcript-exon information'
+        )
+        annotate_parser.add_argument(
+            '--per_base_coverage',
+            help='per base coverage data (output from mosdepth)'
+        )
+
+    def add_calculate_sample_stats(self):
+        """
+        _summary_
+        """
+        stats_parser = self.subparsers.add_parser(
+            'calculate_sample_stats',
+            help='generate per gene and per exon coverage stats'
+        )
+        stats_parser.add_argument(
+            '--annotated_bed',
+            help='per base coverage bed file from annotated_bed'
+        )
+        stats_parser.add_argument(
+            '--thresholds', nargs='*',
+            default=[10, 20, 30, 50, 100],
+            help=(
+                'threshold values to calculate coverage for as comma\
+                seperated integers (default: 10, 20, 30, 50, 100).'
+            )
+        )
+        stats_parser.add_argument(
+            '--hsmetrics', nargs='?', required=False,
+            help=(
+                'Optional hsmetrics file, needed for generating run stats. '
+                'If given metrics will be written to first lines of exon stats '
+                'file.'
+            )
+        )
+        stats_parser.add_argument(
+            '--build', nargs='?', required=False,
+            help='Optional text file with build number used for alignment.'
+        )
